Refactor UI structure and styles
