+++ src/App.jsx
@@ -1,33 +1,33 @@
 import { useEffect } from "react";
 import { Route, Routes } from "react-router"; // убедитесь, что у вас react-router-dom (v6)
 import Lenis from "@studio-freight/lenis";          // <-- правильный импорт!
-import Preloader from "./UI/preloader";
+import Preloader from "./UI/Preloader/Preloader";
 import MainPage from "./pages/MainPage";
 import MainPagesLayout from "./pages/layouts/MainPageLayouts";
-import "./App.css"
+
 function startNoise() {
   const canvas = document.getElementById('noiseCanvas');
   const ctx = canvas.getContext('2d');
-  canvas.width = window.innerWidth;
-  canvas.height = window.innerHeight;
 
-  const imageData = ctx.createImageData(canvas.width, canvas.height);
+  const SIZE = 200; // render small canvas and scale via CSS
+  canvas.width = SIZE;
+  canvas.height = SIZE;
+  canvas.style.width = '100%';
+  canvas.style.height = '100%';
+
+  const imageData = ctx.createImageData(SIZE, SIZE);
   const buffer32 = new Uint32Array(imageData.data.buffer);
 
   function noiseFrame() {
+    crypto.getRandomValues(buffer32);
     for (let i = 0; i < buffer32.length; i++) {
-      const gray = Math.random() * 255 | 0;
-      buffer32[i] = (255 << 24) | (gray << 16) | (gray << 8) | gray;
+      buffer32[i] |= 0xff000000; // set alpha channel
     }
     ctx.putImageData(imageData, 0, 0);
+    requestAnimationFrame(noiseFrame);
   }
 
-  function loop() {
-    noiseFrame();
-    requestAnimationFrame(loop);
-  }
-
-  loop();
+  noiseFrame();
 }
 function App() {
   useEffect(() => {
