+++ src/UI/About/About.jsx
@@ -2,6 +2,7 @@
 import React, { useRef, useEffect } from 'react';
 import { gsap } from 'gsap';
 import { ScrollTrigger } from 'gsap/ScrollTrigger';
+import './About.css';
 
 gsap.registerPlugin(ScrollTrigger);
