+++ src/UI/Magnifier/Magnifier.css
@@ -1,0 +1,12 @@
+.magnifier {
+  position: absolute;
+  pointer-events: none;
+  border-radius: 50%;
+  border: 2px solid #000;
+  box-shadow: 0 0 8px rgba(0, 0, 0, 0.3);
+  background-repeat: no-repeat;
+  display: none;
+  z-index: 10000;
+  bottom: 4%;
+  right: 8%;
+}
