+++ src/UI/Preloader/Preloader.css
@@ -1,0 +1,79 @@
+.preloader {
+  position: fixed;
+  z-index: 100;
+  background-color: white;
+  width: 100%;
+  height: 100vh;
+  top: 0;
+  left: 0;
+  display: flex;
+  flex-direction: row;
+  gap: 4%;
+  justify-content: space-between;
+  align-items: center;
+}
+
+.preloader__background {
+  background-color: black;
+  position: absolute;
+  width: 2000vw;
+  height: 100vh;
+  top: 0;
+  right: 0;
+  z-index: -1;
+  display: flex;
+}
+
+.line--type1,
+.line--type2 {
+  width: 10%;
+  height: 100vh;
+  background-color: #0c0c0c;
+}
+
+.line--type2 {
+  background-color: #000000;
+}
+
+.preloader__logo--string {
+  font-size: max(100px, 20vw);
+  line-height: max(40px, 15vw);
+}
+
+.preloader__desc {
+  font-size: max(20px, 3vw);
+  text-align: right;
+}
+
+.preloader::after {
+  content: "";
+  width: 100vw;
+  height: 80vh;
+  margin-top: 6vh;
+  position: absolute;
+  top: 0;
+  right: 0;
+  border-bottom: 2vw dashed black;
+@media screen and (max-width: 500px) {
+  .preloader {
+    flex-direction: column;
+    justify-content: center;
+  }
+
+  .preloader__logo--string {
+    line-height: max(60px, 20vw);
+  }
+
+  .face {
+    height: 80vh;
+    width: 140vw;
+    left: -30vw;
+    top: 5vh
+  }
+
+  .preloader::after {
+    margin-top: 10vh;
+  }
+
+}
+
