--- src/UI/link__button.jsx
@@ -1,3 +1,0 @@
-function LinkButton(params) {
-    
-}
