+++ src/main.jsx
@@ -1,7 +1,7 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import { BrowserRouter } from "react-router";
-import './index.css'
+import './global.css'
 import App from './App.jsx'
 
 createRoot(document.getElementById('root')).render(
