+++ src/pages/layouts/MainPageLayouts.jsx
@@ -1,5 +1,5 @@
 import { Outlet } from "react-router";
-import Header from "../../UI/header";
+import Header from "../../UI/Header/Header";
 function MainPagesLayout() {
     return (
         <>
